+++ lib/trmnl_preview/cmd/usage.rb
@@ -5,6 +5,7 @@ Usage:
 
 Commands (-h for command-specific help):
 
-  serve     Start the TRMNL Preview server
   build     Generate static HTML files
+  serve     Start the TRMNL Preview server
+  version   Print the version number
 USAGE