+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
-## Next version
+## 0.3.2
 
+- Add bitmap rendering
 - Add TRMNL's [custom plugin filters](https://help.usetrmnl.com/en/articles/10347358-custom-plugin-filters)
 - Add support for user-supplied custom filters